+++ async-http-client-backend/cats/src/test/scala/com/softwaremill/sttp/asynchttpclient/cats/AsyncHttpClientCatsHttpTest.scala
@@ -1,0 +1,12 @@
+package com.softwaremill.sttp.asynchttpclient.cats
+
+import cats.effect.IO
+import com.softwaremill.sttp.SttpBackend
+import com.softwaremill.sttp.impl.cats.convertCatsIOToFuture
+import com.softwaremill.sttp.testing.{ConvertToFuture, HttpTest}
+
+class AsyncHttpClientCatsHttpTest extends HttpTest[IO] {
+
+  override implicit val backend: SttpBackend[IO, Nothing] = AsyncHttpClientCatsBackend()
+  override implicit val convertToFuture: ConvertToFuture[IO] = convertCatsIOToFuture
+}
