+++ async-http-client-backend/scalaz/src/test/scala/com/softwaremill/sttp/asynchttpclient/scalaz/AsyncHttpClientScalazHttpTest.scala
@@ -1,0 +1,13 @@
+package com.softwaremill.sttp.asynchttpclient.scalaz
+
+import com.softwaremill.sttp.SttpBackend
+import com.softwaremill.sttp.impl.scalaz.convertScalazTaskToFuture
+import com.softwaremill.sttp.testing.{ConvertToFuture, HttpTest}
+
+import scalaz.concurrent.Task
+
+class AsyncHttpClientScalazHttpTest extends HttpTest[Task] {
+
+  override implicit val backend: SttpBackend[Task, Nothing] = AsyncHttpClientScalazBackend()
+  override implicit val convertToFuture: ConvertToFuture[Task] = convertScalazTaskToFuture
+}
