+++ implementations/cats/src/test/scala/com/softwaremill/sttp/impl/cats/package.scala
@@ -1,7 +1,7 @@
 package com.softwaremill.sttp.impl
 
 import _root_.cats.effect.IO
-import com.softwaremill.sttp.testing.streaming.ConvertToFuture
+import com.softwaremill.sttp.testing.ConvertToFuture
 
 import scala.concurrent.Future
