+++ tests/src/test/scala/com/softwaremill/sttp/EvalScala.scala
@@ -1,0 +1,11 @@
+package com.softwaremill.sttp
+
+object EvalScala {
+  import scala.tools.reflect.ToolBox
+
+  def apply(code: String): Any = {
+    val m = scala.reflect.runtime.currentMirror
+    val tb = m.mkToolBox()
+    tb.eval(tb.parse(code))
+  }
+}
